add overwrite confirmation

+++ src/controller.py
@@ -3,6 +3,7 @@ from wav_io import ZsndWavReader, ZsndWavWriter
 from util import ZsndLogMixin
 
 from rich.progress import Progress
+import typer
 from i18n import t as _
 import io
 import os
@@ -13,14 +14,11 @@ class StripZsndController(StrippingEventListener, ZsndLogMixin):
     Controller class should take care of environment‑specific concerns (data stores, UX and so on),
     so that Service class can focus exclusively on core business logic and become more testable.
     '''
-    def __init__(self):
-        self.service = StripZsndService(self)
-
     def _do_strip(self, reader: ZsndWavReader, writer, min_duration, threshold, detect_only):
         with Progress() as progress:
             self._progress = progress
             self._progress_task = progress.add_task(_('app.processing'), total=reader.count_frames())
-            self.service.strip(reader, writer, min_duration, threshold, detect_only)
+            StripZsndService().strip(reader, writer, min_duration, threshold, detect_only, self)
 
     @override
     def on_update_progress(self, num_processed_frames, num_total_frames):
@@ -28,7 +26,8 @@ class StripZsndController(StrippingEventListener, ZsndLogMixin):
         assert self._progress_task is not None
         self._progress.update(self._progress_task, completed=num_processed_frames, total=num_total_frames)
 
-    def strip(self, input: str, writer: str|None, min_duration: int, threshold: float, detect_only: bool):
+    def strip(self, input: str, output: str|None, force_overwrite: bool, 
+            min_duration: int, threshold: float, detect_only: bool):
         out_file: io.BufferedWriter|None = None
         writer = None
         in_file, reader = self._create_reader(input)
@@ -37,13 +36,15 @@ class StripZsndController(StrippingEventListener, ZsndLogMixin):
         try:
             if not detect_only:
                 output_base, ext = os.path.splitext(input)
-                output_path = writer or f'{output_base}-fix{ext}'
-                out_file, writer = self._create_writer(output_path, reader)
+                output_path = output or f'{output_base}-fix{ext}'
+                out_file, writer = self._create_writer(output_path, reader, force_overwrite)
                 if writer is None:
                     return 1
 
             return self._do_strip(reader, writer, min_duration, threshold, detect_only)
 
+        except typer.Exit:
+            raise
         except Exception as exc:
             logger = self.get_logger()
             logger.error(str(exc))
@@ -66,21 +67,28 @@ class StripZsndController(StrippingEventListener, ZsndLogMixin):
                 raise
         except Exception as exc:
             logger = self.get_logger()
-            logger.error(_('zsnd.input_file_cannot_be_opened'), {'f': path, 'exc': str(exc)})
+            logger.error(_('app.input_file_cannot_be_opened'), {'f': path, 'exc': str(exc)})
             logger.debug('', exc_info=True)
             return (None, None)
 
-    def _create_writer(self, path: str, reader: ZsndWavReader) -> tuple[io.BufferedIOBase|None, ZsndWavWriter|None]:
+    def _create_writer(self, path: str, reader: ZsndWavReader, force_overwrite: bool) \
+            -> tuple[io.BufferedIOBase|None, ZsndWavWriter|None]:
         try:
+            if os.path.exists(path) and not force_overwrite:
+                if not typer.confirm(_('app.confirm_overwrite_output_file') % (path)):
+                    raise typer.Exit(0)
             outf = io.open(path, 'wb')
             try:
-                return (outf, ZsndWavWriter(outf, reader))
+                bytes_per_sample = reader.get_wave_format().get_bytes_per_sample()
+                return (outf, ZsndWavWriter(outf, bytes_per_sample, reader.get_sample_rate()))
             except BaseException as exc:
                 outf.close()
                 raise
+        except typer.Exit:
+            raise
         except Exception as exc:
             logger = self.get_logger()
-            logger.error(_('zsnd.output_file_cannot_be_opened'),
+            logger.error(_('app.output_file_cannot_be_opened'),
                     {'f': path, 'exc': str(exc)})
             logger.debug('', exc_info=True)
             return (None, None)
